requirements.txt
backports.functools-lru-cache==1.5
certifi==2019.6.16
chardet==3.0.4
cheroot==6.5.6
CherryPy==17.4.2
contextlib2==0.5.5
idna==2.8
jaraco.functools==2.0
more-itertools==5.0.0
numpy==1.16.4
pandas==0.24.2
portend==2.5
python-dateutil==2.8.0
pytz==2019.2
redis==3.3.8
requests==2.22.0
six==1.12.0
tempora==1.14.1
urllib3==1.25.3
zc.lockfile==2.0
